ipfixprobe - introduce process FLOW HASH plugin
