+++ src/plugins/process/CMakeLists.txt
@@ -2,3 +2,4 @@ add_subdirectory(basicplus)
 add_subdirectory(bstats)
 add_subdirectory(icmp)
 add_subdirectory(vlan)
+add_subdirectory(flowHash)
